requirements.txt
attrs==22.1.0
bibliograph.core==1.0.3
bibliograph.parsing==1.0.1
blis==0.9.1
catalogue==2.0.8
certifi==2022.6.15.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.1
cryptography==38.0.1
cymem==2.0.6
docx==0.2.4
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
et-xmlfile==1.1.0
filetype==1.1.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.8
nltk==3.7
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
pathy==0.6.2
pdfminer.six==20220524
pdfplumber==0.7.4
phonenumbers==8.12.55
Pillow==9.2.0
preshed==3.0.7
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
pyparsing==3.0.9
PyPDF2==2.10.7
pyresparser==1.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
resume-parser==0.8.4
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stemming==1.0.1
thinc==8.1.1
tika==1.24
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
Wand==0.6.10
wasabi==0.10.1
